+++ pkg/runner/step_parser.go
@@ -80,7 +80,7 @@ func (p *Parser) parseStepResource(stepRef *proto.Step_Reference) (StepResource,
 		return NewGitStepResource(stepRef.Url, stepRef.Version, stepRef.Path, stepRef.Filename), nil
 
 	case proto.StepReferenceProtocol_oci:
-		return NewOCIStepResource(stepRef.Url, stepRef.Version, stepRef.Path, stepRef.Filename), nil
+		return NewOCIStepResource(stepRef.Registry, stepRef.Repository, stepRef.Tag, stepRef.Path, stepRef.Filename), nil
 
 	case proto.StepReferenceProtocol_builtin:
 		return NewBuiltInStepResource(stepRef.Path, stepRef.Filename), nil